verify-pairs/requirements.txt
javalang==0.11.0
requests==2.18.1
typing==3.6.1
javalang==0.11.0
python-Levenshtein==0.12.0
